Fix incorrect theta values when atan returns the wrong quadrant (#2)
# Summary
atan is defined to return a value between -90 and 90 degrees, thus when converting values in the second or third quadrant the angle would be pointing the wrong direction.

Offset the angle by 180 degrees when working in the second or third quadrants (with negative x values) to ensure the angle is correct. Also convert negative angles to positive angles to line up with how I like to think.

## Test Plan
Tests included which test all four quadrants to ensure the returned angle is correct.